/heatbeat again complains about missing cookies [fixes #143]
the missing-cookie-complaint was replaced by
if-there-is-no-cookie-we'll-create-one. that might work well in other
places, but not if we actually want to check if the user accepts
cookies.

we need that check, because everything relies on cookies. we do not have
auth or something to give the users their archives. workspaces are only
assigned by cookie values.